requirements.txt
code128==0.3
icmplib==3.0.4
Jinja2==3.1.3
Pillow==10.4.0
pyodbc==5.0.1
python_dateutil==2.8.2
pytz==2023.3.post1
Requests==2.32.3
setup==1.0
twilio==8.10.2
